vector and AWNCC max_roh2

+++ Typedefine.hpp
@@ -66,15 +66,15 @@ typedef struct tagF3DPoint
 
 typedef struct tagD2DPoint
 {
-	float m_X;
-	float m_Y;
+	double m_X;
+	double m_Y;
 } D2DPOINT;
 
 typedef struct tagD3DPoint
 {
-	float m_X;
-	float m_Y;
-	float m_Z;
+	double m_X;
+	double m_Y;
+	double m_Z;
     uint8 flag;
 } D3DPOINT;
 
@@ -114,14 +114,14 @@ typedef struct tagNCCflag
 
 typedef struct tagNCCresult
 {
-	float result0; //first peak roh
-	float result1; //second peak roh
+	double result0; //first peak roh
+	double result1; //second peak roh
 	float result2; //first peak height
 	float result3; //second peak height
     float GNCC;
     //float *GNCC_multi;
 	int result4; //peak count
-    float max_WNCC;
+    double max_WNCC;
     int max_WNCC_pos;
 	
     int minHeight;
@@ -136,9 +136,9 @@ typedef struct UpdateGrid{
 	int maxHeight;
 	
 	float Height; //after blunder detection
-	float roh;
-	float ortho_ncc[MaxNCC];
-    float Mean_ortho_ncc;
+	double roh;
+	double ortho_ncc[MaxNCC];
+    double Mean_ortho_ncc;
 
     uint8 Matched_flag;
 	uint8 anchor_flag;
